refactor launcher
