+++ packages/core/engine/main.spec.ts
@@ -1,4 +1,4 @@
-import engine from '.';
+import { engine } from './main';
 
 describe('kites init', () => {
   it('should init a new kites instance without options!', async () => {
