+++ packages/core/engine/main.ts
@@ -1,0 +1,23 @@
+import * as path from 'path';
+import { IKitesOptions, KitesInstance } from './kites';
+
+/**
+ * Return a new kites instance
+ * @param options
+ */
+export function engine(options?: IKitesOptions | boolean) {
+  if (typeof options === 'boolean') {
+    options = {
+      discover: options
+    };
+  }
+
+  // create kites instance
+  let parent = module.parent || module;
+  let opts = Object.assign({
+    parentModuleDirectory: path.dirname(parent.filename)
+  }, options);
+
+  // init a new kites
+  return new KitesInstance(opts);
+}
