+++ packages/core/index.ts
@@ -1,6 +1,3 @@
-import * as path from 'path';
-import {IKitesOptions, KitesInstance} from './engine/kites';
-
 export {
   EventCollectionEmitter,
   ICollectionItem,
@@ -8,23 +5,4 @@ export {
 export { IKites, KitesInstance, IKitesOptions } from './engine/kites';
 export { KitesExtension, ExtensionDefinition, ExtensionOptions } from './extensions/extensions';
 
-/**
- * Return a new kites instance
- * @param options
- */
-export default function engine(options?: IKitesOptions | boolean) {
-    if (typeof options === 'boolean') {
-        options = {
-            discover: options
-        };
-    }
-
-    // create kites instance
-    let parent = module.parent || module;
-    let opts = Object.assign({
-        parentModuleDirectory: path.dirname(parent.filename)
-    }, options);
-
-    // init a new kites
-    return new KitesInstance(opts);
-}
+export * from './engine/main';
