+++ tslint.json
@@ -49,7 +49,8 @@
     "indent": [
       true,
       "spaces"
-    ]
+    ],
+    "space-before-function-paren": false
   },
   "rulesDirectory": []
 }
